+++ src/sll/api/sys_api.c
@@ -20,9 +20,6 @@
 
 static sll_array_length_t _sys_argc=0;
 static sll_string_t* _sys_argv=NULL;
-static sll_string_t _sys_e=SLL_INIT_STRING_STRUCT;
-static sll_string_t _sys_l=SLL_INIT_STRING_STRUCT;
-static sll_string_t _sys_p=SLL_INIT_STRING_STRUCT;
 static library_t** _sys_lh=NULL;
 static sll_array_length_t _sys_lhl=0;
 static char _sys_end=0;
@@ -36,18 +33,6 @@ static void _sys_free_data(void){
 		}
 		sll_deallocate(_sys_argv);
 		_sys_argv=NULL;
-	}
-	if (_sys_e.l){
-		sll_free_string(&_sys_e);
-		_sys_e.l=0;
-	}
-	if (_sys_l.l){
-		sll_free_string(&_sys_l);
-		_sys_l.l=0;
-	}
-	if (_sys_p.l){
-		sll_free_string(&_sys_p);
-		_sys_p.l=0;
 	}
 	if (_sys_lhl){
 		while (_sys_lhl){
@@ -138,44 +123,21 @@ __API_FUNC(sys_get_env){
 
 
 __API_FUNC(sys_get_executable){
-	if (!_sys_e.l){
-		sll_char_t bf[SLL_API_MAX_FILE_PATH_LENGTH];
-		sll_string_length_t bfl=sll_platform_get_executable_file_path(bf,SLL_API_MAX_FILE_PATH_LENGTH);
-		sll_string_from_pointer_length(bf,bfl,&_sys_e);
-		if (!_sys_end){
-			sll_register_cleanup(_sys_free_data);
-			_sys_end=1;
-		}
-	}
-	sll_string_clone(&_sys_e,out);
+	sll_char_t bf[SLL_API_MAX_FILE_PATH_LENGTH];
+	sll_string_from_pointer_length(bf,sll_platform_get_executable_file_path(bf,SLL_API_MAX_FILE_PATH_LENGTH),out);
 }
 
 
 
 __API_FUNC(sys_get_library){
-	if (!_sys_l.l){
-		sll_char_t bf[SLL_API_MAX_FILE_PATH_LENGTH];
-		sll_string_length_t bfl=sll_platform_get_library_file_path(bf,SLL_API_MAX_FILE_PATH_LENGTH);
-		sll_string_from_pointer_length(bf,bfl,&_sys_l);
-		if (!_sys_end){
-			sll_register_cleanup(_sys_free_data);
-			_sys_end=1;
-		}
-	}
-	sll_string_clone(&_sys_l,out);
+	sll_char_t bf[SLL_API_MAX_FILE_PATH_LENGTH];
+	sll_string_from_pointer_length(bf,sll_platform_get_library_file_path(bf,SLL_API_MAX_FILE_PATH_LENGTH),out);
 }
 
 
 
 __API_FUNC(sys_get_platform){
-	if (!_sys_p.l){
-		sll_string_from_pointer(sll_platform_string,&_sys_p);
-		if (!_sys_end){
-			sll_register_cleanup(_sys_free_data);
-			_sys_end=1;
-		}
-	}
-	sll_string_clone(&_sys_p,out);
+	sll_string_from_pointer(sll_platform_string,out);
 }