+++ src/sll/include/sll/platform.h
@@ -343,6 +343,20 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_pid_t sll_platform_get_pid(void);
 
 /**
  * \flags check_output func
+ * \name sll_platform_get_temporary_file_path
+ * \group platform
+ * \subgroup platform-path
+ * \desc Docs!
+ * \arg sll_char_t* o
+ * \arg sll_string_length_t ol
+ * \ret sll_string_length_t
+ */
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_string_length_t sll_platform_get_temporary_file_path(sll_char_t* o,sll_string_length_t ol);
+
+
+
+/**
+ * \flags check_output func
  * \name sll_platform_list_directory
  * \group platform
  * \subgroup platform-path